Remove layer of abstraction
